requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
dj-database-url==0.4.2
Django==1.11.5
django-filter==1.0.4
django-oauth-toolkit==1.0.0
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.4
psycopg2==2.7.3.1
PyJWT==1.5.3
pytz==2017.2
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
